move CUIRect class to separate files
