+++ src/game/client/ui.cpp
@@ -13,8 +13,6 @@
  UI
 *********************************************************/
 
-IGraphics *CUIRect::m_pGraphics = 0;
-
 const float CUI::ms_ListheaderHeight = 17.0f;
 const float CUI::ms_FontmodHeight = 0.8f;
 
@@ -202,175 +200,6 @@ void CUI::UpdateClipping()
 	{
 		Graphics()->ClipDisable();
 	}
-}
-
-void CUIRect::HSplitMid(CUIRect *pTop, CUIRect *pBottom, float Spacing) const
-{
-	CUIRect r = *this;
-	const float Cut = r.h/2;
-	const float HalfSpacing = Spacing/2;
-
-	if(pTop)
-	{
-		pTop->x = r.x;
-		pTop->y = r.y;
-		pTop->w = r.w;
-		pTop->h = Cut - HalfSpacing;
-	}
-
-	if(pBottom)
-	{
-		pBottom->x = r.x;
-		pBottom->y = r.y + Cut + HalfSpacing;
-		pBottom->w = r.w;
-		pBottom->h = r.h - Cut - HalfSpacing;
-	}
-}
-
-void CUIRect::HSplitTop(float Cut, CUIRect *pTop, CUIRect *pBottom) const
-{
-	CUIRect r = *this;
-
-	if (pTop)
-	{
-		pTop->x = r.x;
-		pTop->y = r.y;
-		pTop->w = r.w;
-		pTop->h = Cut;
-	}
-
-	if (pBottom)
-	{
-		pBottom->x = r.x;
-		pBottom->y = r.y + Cut;
-		pBottom->w = r.w;
-		pBottom->h = r.h - Cut;
-	}
-}
-
-void CUIRect::HSplitBottom(float Cut, CUIRect *pTop, CUIRect *pBottom) const
-{
-	CUIRect r = *this;
-
-	if (pTop)
-	{
-		pTop->x = r.x;
-		pTop->y = r.y;
-		pTop->w = r.w;
-		pTop->h = r.h - Cut;
-	}
-
-	if (pBottom)
-	{
-		pBottom->x = r.x;
-		pBottom->y = r.y + r.h - Cut;
-		pBottom->w = r.w;
-		pBottom->h = Cut;
-	}
-}
-
-
-void CUIRect::VSplitMid(CUIRect *pLeft, CUIRect *pRight, float Spacing) const
-{
-	CUIRect r = *this;
-	const float Cut = r.w/2;
-	const float HalfSpacing = Spacing/2;
-
-	if (pLeft)
-	{
-		pLeft->x = r.x;
-		pLeft->y = r.y;
-		pLeft->w = Cut - HalfSpacing;
-		pLeft->h = r.h;
-	}
-
-	if (pRight)
-	{
-		pRight->x = r.x + Cut + HalfSpacing;
-		pRight->y = r.y;
-		pRight->w = r.w - Cut - HalfSpacing;
-		pRight->h = r.h;
-	}
-}
-
-void CUIRect::VSplitLeft(float Cut, CUIRect *pLeft, CUIRect *pRight) const
-{
-	CUIRect r = *this;
-
-	if (pLeft)
-	{
-		pLeft->x = r.x;
-		pLeft->y = r.y;
-		pLeft->w = Cut;
-		pLeft->h = r.h;
-	}
-
-	if (pRight)
-	{
-		pRight->x = r.x + Cut;
-		pRight->y = r.y;
-		pRight->w = r.w - Cut;
-		pRight->h = r.h;
-	}
-}
-
-void CUIRect::VSplitRight(float Cut, CUIRect *pLeft, CUIRect *pRight) const
-{
-	CUIRect r = *this;
-
-	if (pLeft)
-	{
-		pLeft->x = r.x;
-		pLeft->y = r.y;
-		pLeft->w = r.w - Cut;
-		pLeft->h = r.h;
-	}
-
-	if (pRight)
-	{
-		pRight->x = r.x + r.w - Cut;
-		pRight->y = r.y;
-		pRight->w = Cut;
-		pRight->h = r.h;
-	}
-}
-
-void CUIRect::Margin(float Cut, CUIRect *pOtherRect) const
-{
-	CUIRect r = *this;
-
-	pOtherRect->x = r.x + Cut;
-	pOtherRect->y = r.y + Cut;
-	pOtherRect->w = r.w - 2*Cut;
-	pOtherRect->h = r.h - 2*Cut;
-}
-
-void CUIRect::VMargin(float Cut, CUIRect *pOtherRect) const
-{
-	CUIRect r = *this;
-
-	pOtherRect->x = r.x + Cut;
-	pOtherRect->y = r.y;
-	pOtherRect->w = r.w - 2*Cut;
-	pOtherRect->h = r.h;
-}
-
-void CUIRect::HMargin(float Cut, CUIRect *pOtherRect) const
-{
-	CUIRect r = *this;
-
-	pOtherRect->x = r.x;
-	pOtherRect->y = r.y + Cut;
-	pOtherRect->w = r.w;
-	pOtherRect->h = r.h - 2*Cut;
-}
-
-bool CUIRect::Inside(float x, float y) const
-{
-	return x >= this->x
-		&& x < this->x + this->w
-		&& y >= this->y
-		&& y < this->y + this->h;
 }
 
 bool CUI::DoButtonLogic(const void *pID, const CUIRect *pRect, int Button)
@@ -853,279 +682,6 @@ void CUI::DoScrollbarOptionLabeled(void *pID, int *pOption, const CUIRect *pRect
 		Value = (Value + 1) % Num;
 
 	*pOption = clamp(Value, 0, Max);
-}
-
-void CUIRect::Draw(const vec4 &Color, float Rounding, int Corners) const
-{
-	m_pGraphics->TextureClear();
-
-	// TODO: FIX US
-	m_pGraphics->QuadsBegin();
-	m_pGraphics->SetColor(Color.r*Color.a, Color.g*Color.a, Color.b*Color.a, Color.a);
-
-	const float r = Rounding;
-
-	IGraphics::CFreeformItem a_FreeformItems[NUM_ROUND_CORNER_SEGMENTS/2*8];
-	int NumFreeformItems = 0;
-	for(int i = 0; i < NUM_ROUND_CORNER_SEGMENTS; i += 2)
-	{
-		float a1 = i/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float a2 = (i+1)/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float a3 = (i+2)/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float Ca1 = cosf(a1);
-		float Ca2 = cosf(a2);
-		float Ca3 = cosf(a3);
-		float Sa1 = sinf(a1);
-		float Sa2 = sinf(a2);
-		float Sa3 = sinf(a3);
-
-		if(Corners&CORNER_TL)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+r, y+r,
-				x+(1-Ca1)*r, y+(1-Sa1)*r,
-				x+(1-Ca3)*r, y+(1-Sa3)*r,
-				x+(1-Ca2)*r, y+(1-Sa2)*r);
-
-		if(Corners&CORNER_TR)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+w-r, y+r,
-				x+w-r+Ca1*r, y+(1-Sa1)*r,
-				x+w-r+Ca3*r, y+(1-Sa3)*r,
-				x+w-r+Ca2*r, y+(1-Sa2)*r);
-
-		if(Corners&CORNER_BL)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+r, y+h-r,
-				x+(1-Ca1)*r, y+h-r+Sa1*r,
-				x+(1-Ca3)*r, y+h-r+Sa3*r,
-				x+(1-Ca2)*r, y+h-r+Sa2*r);
-
-		if(Corners&CORNER_BR)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+w-r, y+h-r,
-				x+w-r+Ca1*r, y+h-r+Sa1*r,
-				x+w-r+Ca3*r, y+h-r+Sa3*r,
-				x+w-r+Ca2*r, y+h-r+Sa2*r);
-
-		if(Corners&CORNER_ITL)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x, y,
-				x+(1-Ca1)*r, y-r+Sa1*r,
-				x+(1-Ca3)*r, y-r+Sa3*r,
-				x+(1-Ca2)*r, y-r+Sa2*r);
-
-		if(Corners&CORNER_ITR)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+w, y,
-				x+w-r+Ca1*r, y-r+Sa1*r,
-				x+w-r+Ca3*r, y-r+Sa3*r,
-				x+w-r+Ca2*r, y-r+Sa2*r);
-
-		if(Corners&CORNER_IBL)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x, y+h,
-				x+(1-Ca1)*r, y+h+(1-Sa1)*r,
-				x+(1-Ca3)*r, y+h+(1-Sa3)*r,
-				x+(1-Ca2)*r, y+h+(1-Sa2)*r);
-
-		if(Corners&CORNER_IBR)
-			a_FreeformItems[NumFreeformItems++] = IGraphics::CFreeformItem(
-				x+w, y+h,
-				x+w-r+Ca1*r, y+h+(1-Sa1)*r,
-				x+w-r+Ca3*r, y+h+(1-Sa3)*r,
-				x+w-r+Ca2*r, y+h+(1-Sa2)*r);
-	}
-	m_pGraphics->QuadsDrawFreeform(a_FreeformItems, NumFreeformItems);
-
-	IGraphics::CQuadItem a_QuadItems[9];
-	int NumQuadItems = 0;
-	a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+r, y+r, w-r*2, h-r*2); // center
-	a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+r, y, w-r*2, r); // top
-	a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+r, y+h-r, w-r*2, r); // bottom
-	a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x, y+r, r, h-r*2); // left
-	a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+w-r, y+r, r, h-r*2); // right
-
-	if(!(Corners&CORNER_TL)) a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x, y, r, r);
-	if(!(Corners&CORNER_TR)) a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+w, y, -r, r);
-	if(!(Corners&CORNER_BL)) a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x, y+h, r, -r);
-	if(!(Corners&CORNER_BR)) a_QuadItems[NumQuadItems++] = IGraphics::CQuadItem(x+w, y+h, -r, -r);
-
-	m_pGraphics->QuadsDrawTL(a_QuadItems, NumQuadItems);
-	m_pGraphics->QuadsEnd();
-}
-
-void CUIRect::Draw4(const vec4 &ColorTopLeft, const vec4 &ColorTopRight, const vec4 &ColorBottomLeft, const vec4 &ColorBottomRight, float Rounding, int Corners) const
-{
-	m_pGraphics->TextureClear();
-	m_pGraphics->QuadsBegin();
-
-	const float r = Rounding;
-
-	for(int i = 0; i < NUM_ROUND_CORNER_SEGMENTS; i+=2)
-	{
-		float a1 = i/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float a2 = (i+1)/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float a3 = (i+2)/(float)NUM_ROUND_CORNER_SEGMENTS * pi/2;
-		float Ca1 = cosf(a1);
-		float Ca2 = cosf(a2);
-		float Ca3 = cosf(a3);
-		float Sa1 = sinf(a1);
-		float Sa2 = sinf(a2);
-		float Sa3 = sinf(a3);
-
-		if(Corners&CORNER_TL)
-		{
-			m_pGraphics->SetColor(ColorTopLeft);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+r, y+r,
-				x+(1-Ca1)*r, y+(1-Sa1)*r,
-				x+(1-Ca3)*r, y+(1-Sa3)*r,
-				x+(1-Ca2)*r, y+(1-Sa2)*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_TR)
-		{
-			m_pGraphics->SetColor(ColorTopRight);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+w-r, y+r,
-				x+w-r+Ca1*r, y+(1-Sa1)*r,
-				x+w-r+Ca3*r, y+(1-Sa3)*r,
-				x+w-r+Ca2*r, y+(1-Sa2)*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_BL)
-		{
-			m_pGraphics->SetColor(ColorBottomLeft);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+r, y+h-r,
-				x+(1-Ca1)*r, y+h-r+Sa1*r,
-				x+(1-Ca3)*r, y+h-r+Sa3*r,
-				x+(1-Ca2)*r, y+h-r+Sa2*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_BR)
-		{
-			m_pGraphics->SetColor(ColorBottomRight);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+w-r, y+h-r,
-				x+w-r+Ca1*r, y+h-r+Sa1*r,
-				x+w-r+Ca3*r, y+h-r+Sa3*r,
-				x+w-r+Ca2*r, y+h-r+Sa2*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_ITL)
-		{
-			m_pGraphics->SetColor(ColorTopLeft);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x, y,
-				x+(1-Ca1)*r, y-r+Sa1*r,
-				x+(1-Ca3)*r, y-r+Sa3*r,
-				x+(1-Ca2)*r, y-r+Sa2*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_ITR)
-		{
-			m_pGraphics->SetColor(ColorTopRight);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+w, y,
-				x+w-r+Ca1*r, y-r+Sa1*r,
-				x+w-r+Ca3*r, y-r+Sa3*r,
-				x+w-r+Ca2*r, y-r+Sa2*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_IBL)
-		{
-			m_pGraphics->SetColor(ColorBottomLeft);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x, y+h,
-				x+(1-Ca1)*r, y+h+(1-Sa1)*r,
-				x+(1-Ca3)*r, y+h+(1-Sa3)*r,
-				x+(1-Ca2)*r, y+h+(1-Sa2)*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-
-		if(Corners&CORNER_IBR)
-		{
-			m_pGraphics->SetColor(ColorBottomRight);
-			IGraphics::CFreeformItem ItemF = IGraphics::CFreeformItem(
-				x+w, y+h,
-				x+w-r+Ca1*r, y+h+(1-Sa1)*r,
-				x+w-r+Ca3*r, y+h+(1-Sa3)*r,
-				x+w-r+Ca2*r, y+h+(1-Sa2)*r);
-			m_pGraphics->QuadsDrawFreeform(&ItemF, 1);
-		}
-	}
-
-	// center
-	{
-		m_pGraphics->SetColor4(ColorTopLeft, ColorTopRight, ColorBottomLeft, ColorBottomRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+r, y+r, w-r*2, h-r*2);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	// top
-	{
-		m_pGraphics->SetColor4(ColorTopLeft, ColorTopRight, ColorTopLeft, ColorTopRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+r, y, w-r*2, r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	// bottom
-	{
-		m_pGraphics->SetColor4(ColorBottomLeft, ColorBottomRight, ColorBottomLeft, ColorBottomRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+r, y+h-r, w-r*2, r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	// left
-	{
-		m_pGraphics->SetColor4(ColorTopLeft, ColorTopLeft, ColorBottomLeft, ColorBottomLeft);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x, y+r, r, h-r*2);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	// right
-	{
-		m_pGraphics->SetColor4(ColorTopRight, ColorTopRight, ColorBottomRight, ColorBottomRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+w-r, y+r, r, h-r*2);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	if(!(Corners&CORNER_TL))
-	{
-		m_pGraphics->SetColor(ColorTopLeft);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x, y, r, r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	if(!(Corners&CORNER_TR))
-	{
-		m_pGraphics->SetColor(ColorTopRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+w, y, -r, r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	if(!(Corners&CORNER_BL))
-	{
-		m_pGraphics->SetColor(ColorBottomLeft);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x, y+h, r, -r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	if(!(Corners&CORNER_BR))
-	{
-		m_pGraphics->SetColor(ColorBottomRight);
-		IGraphics::CQuadItem ItemQ = IGraphics::CQuadItem(x+w, y+h, -r, -r);
-		m_pGraphics->QuadsDrawTL(&ItemQ, 1);
-	}
-
-	m_pGraphics->QuadsEnd();
 }
 
 float CUI::DrawClientID(float FontSize, vec2 CursorPosition, int ID,
